rolling back/sidestepping several recent landcover color changes for a more systematic area coloring scheme overall.  Specifically:
playground
pitch/track
dog_park

playground and pitch/track are supposed to form a pair of strong and
prominent colors in the urban context, pitch/track essentially as
'playgrounds for adults'.  These are green tones but meant to be
recognizably separate from the other vegetation related green area
colors.

None of these colors is probably ideal but IMO they are much more
appropriate for representing the importance of the features in
questions and the similarity/dissimilarity to other features.