Designing the transit to a low zoom landcover color scheme based on aggregated classes.
In short at z10 this features a four color landcover scheme using

* a builtup area class (medium gray)
* a low vegetation class (bright yellow-green)
* a tall woody vegetation class (darker green)
* a bare ground class (bright brown-gray)

This reduced color scheme is transited to the full color scheme at
z13 in several steps separating farmland, allotments and beaches
at z11 and further separating all non-urban and some urban
classes at z12.

Colors are based on averaging the major contributing colors from
the high zoom levels in perceptual color space and tuning the
results for a good overall balance.

The distinct industrial landcovers (quarry and landfill) are
currently not included in this - not yet sure how to best treat
them.  Same for the mud base color.

All of this is mostly based on a hunch and has not yet been widely
tested.  Especially the transit from low zoom to high zoom is
highly subject to influence from the latitude dependent scale of
the map.