Changing administrative boundary color from purple to a greenish gray (somewhat similar to the German style).
This serves for both freeing the purpose color for other uses
and to reduce prominence of boundary colors and avoid irritating
color mixing due to transparency.

Also the line width stepping of the boundaries with zoom levels is
smoother which also better avoids confusion with railway lines with
the new color at around z9.