Differentiating maritime boundaries without preprocessed data
This is slightly different from the preprocessing approach used
in the low zoom demo since it uses the tagging as maritime rather
than spatial analysis with the coastline data.

Combines relation based rendering for the land boundaries with way
based rendering for the maritime ones which could produce some
artefacts in some cases.