Re-designing water point feature rendering:
Springs, wells and fountains are rendered with the same base design
in different variants and with two different sizes depending on
zoom level.

Hot springs and intermittent springs are differentiated with a
minor symbol variation.

Suplemental tagging with amenity=drinking_water/drinking_water=yes
is indicated with a secondary icon.

Springs connected to waterways are rendered with a similar design
as standalone springs based on contructed geometries and adjusted
to the line signature of the connected waterway.

Also adjust priorities when the same feature as multiple rendered
tags.