Adding rendering of golf course features
This is based on the French style concept (which is also being
used by the German style) but adjusts it in a number of ways:

* the area rendering is integrated into the general landcover
layer.  This is a bit awkward because the database scheme does
not interpret golf tags to imply closed ways being polygons.
* areas are rendered from z14 while lines, points and labels
are shown from z16.
* the base color of golf courses is changed from scrub to
campsite.

This probably requires some further tuning.