Restructuring the style with the following changes:
Modularization of the style layer definitions with a layers.yaml
file and separate layer definitions in layers/, being combined into
a project.mml file with scripts/assemble_project.py.

A new framework for POI symbol and label rendering that defines the
features to be labeled and the styling to be used in symbols-labels.yaml
and script-generates the SQL and MSS code with
scripts/generate_symbols_labels.py.

Redesigning label rendering:

* based on analysis of composite names in name tag
* matching with individual names
* interpreting default_language tags with the help of preprocessed language_regions polygon data
* using different fonts depending on language of label as determined by default_language and tag matching

This commit contains the code changes, followed by an update/reorganization of the SVG symbols.